[containerfiles] add test detstream pipeline
This is a "working" pythonish version of the detstream pipeline. Here is the flow:
- detstream:worker sends fake data (10x10 stem positions) using detstream_producer
- interactem/detstream-aggregator picks these up using zmq sockets and forwards them
- interactem/detstream-assembler assembles the frames into numpy arrays
- interactem/electron-count counts the electrons

Right now, this will not be performant. But shows that we can send the data all the way down the pipeline.

TO test this, you will need to get github.com/swelborn/detstream and build the images locally.