[WIP] multiprocessing, cont
This is a "working" version of the code. However, the throughput is lower than single-threaded.

Here I initially tried to update each metric from each socket, but this was stupid. We just need to update the recv socket on the main thread metrics to get an idea of throughput.

Drop in performance may be for the following reasons:
1. Spending twice as long deserializing the header.
2. Using my laptop with a bunch of processes/threads running...