add pvapy operators
- adds pvayp-ad-sim-server, which simulates area detector data. This is for testing purposes, and hardcodes the env var `EPICS_PVA_SERVER_PORT=11111`.
- adds pvapy-converter, which takes data from a PVA server and converts it into our framework. Note the env vars (`EPICS_PVA_NAME_SERVERS` in particular) need to be set properly.

Note: this was a pain to get working because there is something funky with EPICS + macOS for discovery. I tried grabbing frames from pvapy-ad-sim-server running outside podman (on macos host), and there seems to be some connection (used `PVAPY_EPICS_LOG_LEVEL=0) but the frame is never received. This DOES work on x86 (tested on perlmutter), so we should be able to receive detector frames from some upstream source that runs outside of a container in production...