autogenerate operator-mode script
This will autogenerate all necessary files for server running in "operator" mode.

We could use this method in the future, but this would involve storing the "frontend_user" credentials somewhere so that the websocket can get to it.

One possible flow is:
1. Store sentinel jwt on the fastAPI server.
2. During auth, send back this jwt to frontend before nats.connect() call.
3. Connect with sentinel jwt + external jwt.
4. This will delegate to auth_callout, and check the external jwt.
5. Auth callout will assign to proper account.

Please refer to the issue raised here:
https://github.com/aricart/callout.go/issues/7

NOTE: something is still incorrect with signing key hierarchy in this.