add constraint to pipeline order
Right now, we want to avoid outbound connections from edge. We make sure that we aren't doing something like this:

Operator A assigned to Agent A
Operator B assigned to Agent B
Operator C assigned to Agent A

where the chain is Operator A -> Operator B -> Operator C.

It could be possible, however, to have two connected operators running on the same agent. For example, we oculd have

Operator A Agent A
Operator B Agent A
Operator C Agent B

but if we were to add another operator here (D), we wouldn't want it to be assigned to Agent A. It should be assigned to Agent B.