Add PipelineDeployment model and related schemas
Introduces the PipelineDeployment model to represent individual pipeline executions, including state management and relationships to PipelineRevision.

We introduce this because we need to dissociate the pipeline ID/revision ID from the deployment ID. This allows for multiple pipelines to be running that have the same ID, and will (future) enable parallelization of operators.