add to_thread for image pulls
we want to be run this sync call in a thread, because if we have to
pull a really big image, we run into a problem that the agent status is
no longer being updated and the agent “dies” according to rest of
network. i think this will solve this problem.